% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr.R
\name{is_purrr_map2_fun}
\alias{is_purrr_map2_fun}
\title{Check if a function is a map2 derived function}
\usage{
is_purrr_map2_fun(fun)
}
\arguments{
\item{fun}{function to test.}
}
\description{
Besides the obvious \link{map2} and \code{map2_*} variants,
this also covers functions based off \code{map2}:
\itemize{
\item \link{imap} and \code{imap_*} variants.
\item \link{invoke_map} and \code{[invoke_map_*]} variants.
}
}
