% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Progress.R
\docType{data}
\name{R6_progress}
\alias{R6_progress}
\title{Base Progress bar Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6_progress
}
\description{
This is the base class for all R6 progress bars.
It also doubles as a null progress bar that displays no progress bar,
but allows for checking values.
}
\keyword{datasets}
