% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_progress.R
\name{with_progress}
\alias{with_progress}
\title{Apply a function with progress bars.}
\usage{
with_progress(fun, total,
  label = "{frac} items completed\\t {etr} remaining", ...,
  type = infer_type(.Platform$OS.type))
}
\arguments{
\item{fun}{The function to be apply}

\item{total}{\link{Optional} the total number of items.
When omitted, a search will be performed to guess the
total.}

\item{label}{The label to be displayed on the progress bar.}

\item{...}{Other arguments to progress bar.}
}
\description{
Apply a function with progress bars.
}
