% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{progress_bar}
\alias{progress_bar}
\title{Create a R6 progress bar directly}
\usage{
progress_bar(total, title = "Progress", label = "", ...,
  type = c("none", "win", "tk"))
}
\arguments{
\item{total}{the total number of elements}

\item{title}{the title of the progress bar}

\item{label}{the label to display}

\item{...}{passed on to the specific constructor determined by type.}

\item{type}{the type of progress bar to create as a string, or an
\code{R6ClassGenerator} object for a class that inherits from
the "R6 Progress Base Class".}
}
\description{
Create a R6 progress bar directly
}
