% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_progress.R
\name{with_progress}
\alias{with_progress}
\title{Apply a function with progress bars.}
\usage{
with_progress(fun, total,
  label = "{frac} items completed\\t {etr} remaining", ...,
  type = infer_type(.Platform$OS.type))
}
\arguments{
\item{fun}{The function to be apply}

\item{total}{the total number of elements}

\item{label}{the label to display}

\item{...}{passed on to the specific constructor determined by type.}

\item{type}{the type of progress bar to create as a string, or an
\code{R6ClassGenerator} object for a class that inherits from
the "R6 Progress Base Class".}
}
\description{
Apply a function with progress bars.
}
